pyproject.toml
[project]
name = "movie-name"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.13.4",
    "cloudinary>=1.44.1",
    "fastapi[all]>=0.116.1",
    "httpx>=0.28.1",
    "ollama>=0.5.3",
    "pillow>=11.3.0",
    "playwright>=1.54.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "selenium>=4.35.0",
    "uvicorn>=0.35.0",
    "webdriver-manager>=4.0.2",
]
